build: strong mode

+++ node/tablestore_node/lib/src/interop/js_node_interop.dart
@@ -18,7 +18,7 @@ DateTime dartifyDate(Object jsObject) {
   if (util.hasProperty(jsObject, 'toDateString')) {
     try {
       var date = jsObject as dynamic;
-      return DateTime.fromMillisecondsSinceEpoch(date.getTime());
+      return DateTime.fromMillisecondsSinceEpoch(date.getTime() as int);
     } on NoSuchMethodError {
       // so it's not a JsDate!
       return null;
